separate server websocket creation to websocket.js
to facilitate proxy parameter testing
